refactor(ui): flatten `plugins` dir
- `plugins-list/` dir had exactly one file in it; redundant dir
- `components/` only had 1 real component, `plugin-list`, plus its wrapper routing container
  - 2 files for a folder not really worth and you had to go 3 folders in to get to the one actual component
