+++ ui/src/app/plugins/index.ts
@@ -1,4 +1,4 @@
-import {PluginsContainer} from './components/plugins-container';
+import {PluginsContainer} from './plugins-container';
 
 export default {
     component: PluginsContainer
